main/requirements.txt
absl-py
albumentations
backcall
cachetools
charset-normalizer
click
colorama
cycler
debugpy
decorator
Deprecated
docopt
easydict
entrypoints
et-xmlfile
fonttools
google-api-core
google-auth
google-auth-oauthlib
googleapis-common-protos
grpcio
humanize
idna
imageio
imgviz
importlib-metadata
ipykernel
ipython
ipywidgets
jedi
joblib
jupyter_client
jupyter_core
jupyterlab-widgets
kiwisolver
kornia
labelbox
labelImg
labelme
lazy_loader
littleutils
lxml
Markdown
MarkupSafe
matplotlib
matplotlib-inline
natsort
nest-asyncio
networkx
nibabel
numpy
oauthlib
opencv-python
opencv-python-headless
openpyxl
outdated
packaging
pandas
pandas-flavor
parso
patsy
pickleshare
Pillow
pingouin
pipreqs
prompt-toolkit
protobuf
psutil
pyasn1
pyasn1-modules
pydantic
pydicom
Pygments
pykwalify
pyparsing
PyQt5
pyqt5-plugins
PyQt5-Qt5
PyQt5-sip
pyqt5-tools
pyradiomics
python-dateutil
python-dotenv
pytz
PyWavelets
pywin32
PyYAML
pyzmq
qt5-applications
qt5-tools
QtPy
qudida
requests
requests-oauthlib
rsa
ruamel.yaml
ruamel.yaml.clib
scikit-image
scikit-learn
scipy
seaborn
SimpleITK
six
statsmodels
tabulate
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
termcolor
threadpoolctl
tifffile
torch
torch-summary
torchaudio
torchio
torchvision
tornado
tqdm
traitlets
typing_extensions
urllib3
wcwidth
Werkzeug
widgetsnbextension
wincertstore
wrapt
xarray
yarg
zipp
